university-quality-os/requirements.txt
# LLM clients
anthropic>=0.45.0
openai>=1.50.0

# Document extraction
pdfplumber>=0.11.0
python-docx>=1.1.0
pytesseract>=0.3.13
Pillow>=10.0.0

# Database + vectors
pgvector>=0.3.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Retrieval (MVP-2)
rank_bm25>=0.2.2

# UI
streamlit>=1.38.0

# Utilities
python-dotenv>=1.0.1
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0

# Testing (for later MVPs)
pytest>=8.3.0
pytest-asyncio>=0.24.0
